Split materialized pg tests
