+++ tests/integration/test_postgresql_replica_database_engine/configs/merge_tree_too_many_parts.xml
@@ -1,0 +1,5 @@
+<clickhouse>
+    <merge_tree>
+        <parts_to_throw_insert>5</parts_to_throw_insert>
+    </merge_tree>
+</clickhouse>
