+++ tests/integration/test_postgresql_replica_database_engine/configs/users.xml
@@ -5,4 +5,11 @@
             <postgresql_fault_injection_probability>0.1</postgresql_fault_injection_probability>
         </default>
     </profiles>
+    <users>
+        <default>
+            <password></password>
+            <profile>default</profile>
+            <named_collection_control>1</named_collection_control>
+        </default>
+    </users>
 </clickhouse>
